refactor(ColumnSettings): merge demo into story
Fold the keyed ColumnSettingsDemo body back into ColumnSettings.story.vue.
The demo was a separate component only to remount useDoctypeMeta (taken
by value) on a doctype switch; a watchEffect that re-calls the
cache-backed composable re-resolves meta reactively in one component, so
the extra file and the :key remount are gone. Same behaviour: switching
doctype reseeds columns from the in_list_view defaults.
